Rich text: use getPasteEventData
