+++ packages/block-editor/src/utils/pasting.js
@@ -3,6 +3,51 @@
  */
 import { getFilesFromDataTransfer } from '@wordpress/dom';
 
+/**
+ * Normalizes a given string of HTML to remove the Windows-specific "Fragment"
+ * comments and any preceding and trailing content.
+ *
+ * @param {string} html the html to be normalized
+ * @return {string} the normalized html
+ */
+function removeWindowsFragments( html ) {
+	const startStr = '<!--StartFragment-->';
+	const startIdx = html.indexOf( startStr );
+	if ( startIdx > -1 ) {
+		html = html.substring( startIdx + startStr.length );
+	} else {
+		// No point looking for EndFragment
+		return html;
+	}
+
+	const endStr = '<!--EndFragment-->';
+	const endIdx = html.indexOf( endStr );
+	if ( endIdx > -1 ) {
+		html = html.substring( 0, endIdx );
+	}
+
+	return html;
+}
+
+/**
+ * Removes the charset meta tag inserted by Chromium.
+ * See:
+ * - https://github.com/WordPress/gutenberg/issues/33585
+ * - https://bugs.chromium.org/p/chromium/issues/detail?id=1264616#c4
+ *
+ * @param {string} html the html to be stripped of the meta tag.
+ * @return {string} the cleaned html
+ */
+function removeCharsetMetaTag( html ) {
+	const metaTag = `<meta charset='utf-8'>`;
+
+	if ( html.startsWith( metaTag ) ) {
+		return html.slice( metaTag.length );
+	}
+
+	return html;
+}
+
 export function getPasteEventData( { clipboardData } ) {
 	let plainText = '';
 	let html = '';
@@ -23,6 +68,12 @@ export function getPasteEventData( { clipboardData } ) {
 			return;
 		}
 	}
+
+	// Remove Windows-specific metadata appended within copied HTML text.
+	html = removeWindowsFragments( html );
+
+	// Strip meta tag.
+	html = removeCharsetMetaTag( html );
 
 	const files = getFilesFromDataTransfer( clipboardData );
